feat: disable the service card when the user is the tutor

+++ client/lib/babysitting-services/create_service_screen.dart
@@ -100,7 +100,7 @@ class _CreateServicePageState extends State<CreateServicePage> {
           initialValue: formData[key].toString(),
           onChanged: (value) {
             setState(() {
-              formData[key] = int.tryParse(value ?? '0') ?? 0;
+              formData[key] = double.tryParse(value ?? '0') ?? 0;
             });
           },
           validator: (value) {
